+++ src/device/graphic/cga/ColorGraphicsAdapter.h
@@ -56,7 +56,7 @@ public:
 
     [[nodiscard]] uint16_t getCurrentRow() const override;
 
-    [[nodiscard]] Util::Address<uint32_t> getAddress();
+    [[nodiscard]] Util::Address<uint64_t> getAddress();
 
 private:
 
@@ -67,7 +67,7 @@ private:
     uint16_t currentColumn = 0;
     uint16_t currentRow = 0;
 
-    Util::Address<uint32_t> cgaMemory;
+    Util::Address<uint64_t> cgaMemory;
     IoPort indexPort = IoPort(0x3d4);
     IoPort dataPort = IoPort(0x3d5);
