+++ src/device/storage/PartitionHandler.cpp
@@ -221,7 +221,7 @@ void PartitionHandler::writePartition(uint8_t partitionNumber, bool active, Syst
 
             // Let next entry of current boot record point to the boot record of the new partition
             auto *newEbr = new uint8_t[device.getSectorSize()];
-            auto newEbrAddress = Util::Address<uint32_t>(newEbr);
+            auto newEbrAddress = Util::Address<uint64_t>(newEbr);
 
             // Create new boot record for the new partition
             newEbrAddress.setRange(0, device.getSectorSize());
@@ -380,7 +380,7 @@ void PartitionHandler::deletePartition(uint8_t partitionNumber) {
 
 void PartitionHandler::createBootRecord(uint32_t sector) {
     auto *mbr = new uint8_t[device.getSectorSize()];
-    auto mbrAddress = Util::Address<uint32_t>(mbr);
+    auto mbrAddress = Util::Address<uint64_t>(mbr);
 
     // Create new empty boot record
     mbrAddress.setRange(0, device.getSectorSize());
