+++ src/kernel/paging/MemoryLayout.h
@@ -49,8 +49,8 @@ public:
             return endAddress - startAddress + 1;
         }
 
-        [[nodiscard]] Util::Address<uint32_t> toAddress() const {
-            return Util::Address<uint32_t>(startAddress);
+        [[nodiscard]] Util::Address<uint64_t> toAddress() const {
+            return Util::Address<uint64_t>(startAddress);
         }
 
         [[nodiscard]] MemoryArea toPhysical() const {
